Moved categoric-specific funcitons into separate extension
